Moves all syscalls into the file src/syscalls.rs
