+++ tests/misc.rs
@@ -23,12 +23,12 @@ extern crate test_utils;
 
 use solana_rbpf::{
     fuzz::fuzz,
+    syscalls::{BpfSyscallString, BpfSyscallU64},
     user_error::UserError,
     verifier::check,
     vm::{Config, DefaultInstructionMeter, EbpfVm, Executable, SyscallObject, SyscallRegistry},
 };
 use std::{fs::File, io::Read};
-use test_utils::{BpfSyscallString, BpfSyscallU64};
 
 // The following two examples have been compiled from C with the following command:
 //
